ai-agent/mcp_agents/requirements.txt
agno==1.1.9
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
attrs==25.1.0
azure-ai-inference==1.0.0b9
azure-core==1.32.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docstring-parser==0.16
frozenlist==1.5.0
gitdb==4.0.12
gitpython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
jiter==0.8.2
markdown-it-py==3.0.0
mcp==1.3.0
mdurl==0.1.2
multidict==6.1.0
openai==1.65.4
propcache==0.3.0
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
